Move Heading to common components
